pyproject.toml
[tool.poetry]
name = "rts"
version = "0.1.0"
description = ""
authors = ["Kirell Benzi <kirell.benzi@epfl.ch>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
jupyterlab = "^3.6.3"
numpy = "^1.23.5"
meilisearch = "^0.26.0"
pandas = "^2.0.0"
orjson = "^3.8.10"
openai-whisper = {git = "https://github.com/openai/whisper.git", rev = "main", markers = "sys_platform != 'darwin'"}
torch = "^2.0.0"
torchvision = "^0.15.1"
torchaudio = "^2.0.1"
tqdm = "^4.65.0"
more-itertools = "^9.1.0"
ffmpeg-python = "^0.2.0"
transformers = "^4.28.1"
tables = "^3.8.0"
matplotlib = "^3.7.1"
av = "^10.0.0"
scenedetect = {extras = ["opencv-headless"], version = "^0.6.1"}
cupy-cuda12x = {version = "^12.0.0", markers = "sys_platform != 'darwin'"}
spacy = {extras = ["cuda-12"], version = "^3.5.2"}
geonamescache = "^1.5.0"
click = "^8.1.3"
pillow = "^9.5.0"
fastapi = {extras = ["all"], version = "^0.95.1"}
img2vec-pytorch = "^1.0.1"
llvmlite = "^0.39.1"
numba = "0.56.4"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
supabase = "^1.0.3"
databases = {extras = ["postgresql"], version = "^0.7.0"}
pytest = "^7.3.1"
alembic = "^1.10.4"
asyncpg = "^0.27.0"
pytest-asyncio = "^0.21.0"
anyio = {extras = ["trio"], version = "^3.6.2"}

[tool.poetry.dev-dependencies]
poethepoet = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.10.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "rts.api.server:start_dev"
prod = "rts.api.server:start_prod"

[tool.poe.tasks]
install-spacy-models = "python -m spacy download fr_core_news_lg"
start-weaviate = "docker compose -f docker/weaviate.yml up -d"
stop-weaviate = "docker compose -f docker/weaviate.yml down"
run-pipeline = "python -m rts.cli pipeline --compute-transcript"
atlas = "python -m rts.cli atlas"
serve = "python -m rts.api.server"
serve-dev = "python -m rts.api.server dev -l rts"
serve-prod = "python -m rts.api.server prod -l rts"
pull-cables = { "shell" = "cables -e 7DZVup -D -d public --skip-backups --no-index" }

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
